requirements.txt
pygame~=2.5.2
python-dotenv~=1.0.1
python-miio~=0.5.12